% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_network.R
\name{coexpression_network}
\alias{coexpression_network}
\title{Creates a gene co-expression network}
\usage{
coexpression_network(
  expr_matrix,
  top_genes = 30,
  method = "pearson",
  threshold = 0.7
)
}
\arguments{
\item{expr_matrix}{Expression matrix (genes x samples).}

\item{top_genes}{Number of genes plus variables to consider (default: 30).}

\item{method}{Correlation method ("pearson" or "spearman").}

\item{threshold}{Absolute correlation threshold to create an arc (default: 0.7).}
}
\value{
igraph object.
}
\description{
Constructs a co-expression graph between the most variable genes.
}
