% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{correct_batch}
\alias{correct_batch}
\title{Correct batch effect with ComBat}
\usage{
correct_batch(expr_data, batch)
}
\arguments{
\item{expr_data}{Expression matrix (genes x samples).}

\item{batch}{Vector indicating the batch of each sample.}
}
\value{
Corrected matrix.
}
\description{
Apply batch effect correction using ComBat.
}
\examples{
data(expr_small)
batch <- c(rep(1, 5), rep(2, 5))
expr_combat <- correct_batch(expr_small, batch)
}
