% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{GO, KEGG, Reactome enrichment analysis}
\usage{
enrichment_analysis(
  gene_list,
  type = "GO",
  organism = "hsa",
  pvalueCutoff = 0.05
)
}
\arguments{
\item{gene_list}{Gene symbol vector or EntrezID.}

\item{type}{Type of analysis: GO, KEGG, Reactome.}

\item{organism}{Organism (default: "hsa" for Homo sapiens).}

\item{pvalueCutoff}{Significance threshold (default: 0.05).}
}
\value{
Object of class enrichResult or similar.
}
\description{
Performs functional enrichment on a list of genes.
}
