% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_auto_hgu95.R
\name{enrichment_auto_hgu95}
\alias{enrichment_auto_hgu95}
\title{Automatic enrichment for probe ID Affymetrix HG-U95 (A/B/C/D/E)}
\usage{
enrichment_auto_hgu95(probe_list, ...)
}
\arguments{
\item{probe_list}{Vector of probe IDs (e.g. "1000_at", "1600_s_at", etc.)}

\item{...}{Other parameters passed to enrichment_analysis (e.g., type = "GO")}
}
\value{
Object enrichResult
}
\description{
Automatically map probe IDs of HG-U95A/B/C/D/E platforms to gene symbols and launch enrichment_analysis().
}
