% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{expr_breast}
\alias{expr_breast}
\title{GSE2508 gene expression subset (breast, example)}
\format{
A numerical matrix (genes x samples)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2508}
}
\usage{
expr_breast
}
\description{
Expression matrix for a random subset of 30 genes and 10 samples from GSE2508 (adipose and breast tissue).
}
\keyword{datasets}
