% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{expr_coculture}
\alias{expr_coculture}
\title{E-MTAB-8632 gene expression subset (co-cultures, example)}
\format{
A numerical matrix (genes x samples)
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-8632/}
}
\usage{
expr_coculture
}
\description{
Expression matrix for a random subset of 30 genes and 10 samples from E-MTAB-8632 (tumor/adipocyte co-cultures).
}
\keyword{datasets}
