% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{expr_colon}
\alias{expr_colon}
\title{Gene expression subset GSE44076 (colon, example)}
\format{
A numerical matrix (genes x samples)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE44076}
}
\usage{
expr_colon
}
\description{
Expression matrix for a random subset of 30 genes and 10 samples from GSE44076 (colon and normal tissue).
}
\keyword{datasets}
