% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{expr_ovary}
\alias{expr_ovary}
\title{GSE40595 gene expression subset (ovary, example)}
\format{
A numerical matrix (genes x samples)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40595}
}
\usage{
expr_ovary
}
\description{
Expression matrix for a random subset of 30 genes and 10 samples from GSE40595 (omental adipose tissue and ovarian tumor).
}
\keyword{datasets}
