% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{expr_small}
\alias{expr_small}
\title{Subset of gene expression data from GSE2508}
\format{
A numerical matrix with 100 genes (rows) and 10 samples (columns).
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2508}
}
\usage{
expr_small
}
\description{
A subset (100 genes x 10 samples) of the GEO dataset GSE2508, which studies gene expression in subcutaneous adipose tissue of breast cancer patients and controls.
}
\keyword{datasets}
