% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{pheno_breast}
\alias{pheno_breast}
\title{Phenotypic table for subset GSE2508 (breast, example)}
\format{
A data frame (samples x phenotypic variables)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2508}
}
\usage{
pheno_breast
}
\description{
Phenotypic table associated with \code{expr_breast}, with information about samples.
}
\keyword{datasets}
