% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{pheno_coculture}
\alias{pheno_coculture}
\title{Phenotypic table for subset E-MTAB-8632 (co-cultures, example)}
\format{
A data frame (samples x phenotypic variables)
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-8632/}
}
\usage{
pheno_coculture
}
\description{
Phenotypic table associated with \code{expr_coculture}, with information about samples.
}
\keyword{datasets}
