% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{pheno_colon}
\alias{pheno_colon}
\title{Phenotypic table for subset GSE44076 (colon, example)}
\format{
A data frame (samples x phenotypic variables)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE44076}
}
\usage{
pheno_colon
}
\description{
Phenotypic table associated with \code{expr_colon}, with information about samples.
}
\keyword{datasets}
