% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{pheno_ovary}
\alias{pheno_ovary}
\title{Phenotypic table for subset GSE40595 (ovary, example)}
\format{
A data frame (samples x phenotypic variables)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40595}
}
\usage{
pheno_ovary
}
\description{
Phenotypic table associated with \code{expr_ovary}, with information about samples.
}
\keyword{datasets}
