% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_network.R
\name{plot_coexpression_network_interactive}
\alias{plot_coexpression_network_interactive}
\title{Display a co-expression network interactively}
\usage{
plot_coexpression_network_interactive(g, node_anno = NULL, palette = rainbow)
}
\arguments{
\item{g}{Object igraph.}

\item{node_anno}{Annotation vector/factor for nodes (optional).}

\item{palette}{Color function for groups (default: rainbow).}
}
\description{
Display a co-expression network interactively
}
\examples{
if (requireNamespace("igraph", quietly = TRUE) && requireNamespace("visNetwork", quietly = TRUE)) {
  set.seed(1)
  expr_matrix <- matrix(rnorm(100), nrow = 10)
  rownames(expr_matrix) <- paste0("Gene", 1:10)
  g <- coexpression_network(expr_matrix, top_genes = 5,threshold = 0.2)
  plot_coexpression_network_interactive(g)
}
}
