% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Display enrichment results (barplot)}
\usage{
plot_enrichment(enrich_result, top = 10)
}
\arguments{
\item{enrich_result}{Object enrichResult (from clusterProfiler or ReactomePA).}

\item{top}{No. of terms to display (default: 10).}
}
\description{
Display enrichment results (barplot)
}
