% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_result.R
\name{plot_enrichment_result}
\alias{plot_enrichment_result}
\title{Display enrichment results (GO, KEGG, Reactome)}
\usage{
plot_enrichment_result(enrich_obj, type = "dotplot", showCategory = 20, ...)
}
\arguments{
\item{enrich_obj}{Object enrichResult (e.g., from enrichGO, enrichKEGG, enrichPathway).}

\item{type}{Type of plot: 'dotplot’, ‘barplot’, ‘cnetplot’, ‘emapplot’, ‘upsetplot’.}

\item{showCategory}{Number of categories to show (default: 20).}

\item{...}{Other parameters for the plotting function.}
}
\value{
A ggplot2 or enrichplot object.
}
\description{
Wrapper for main enrichplot/clusterProfiler plots.
}
\examples{
\dontrun{
# Example (requires enrichResult object)
# plot_enrichment_result(ego, type = "dotplot", showCategory = 10)
}
}
