% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{dotplot_enrichment}
\alias{dotplot_enrichment}
\title{Dotplot of enrichment results}
\usage{
dotplot_enrichment(enrich_result, top = 10)
}
\arguments{
\item{enrich_result}{Object enrichResult.}

\item{top}{No. of terms to display (default: 10).}
}
\description{
Dotplot of enrichment results
}
