% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_enrichr}
\alias{enrichment_enrichr}
\title{Enrichment analysis with enrichR}
\usage{
enrichment_enrichr(
  gene_list,
  databases = c("KEGG_2021_Human", "GO_Biological_Process_2021")
)
}
\arguments{
\item{gene_list}{Vector of gene symbols.}

\item{databases}{Vector of enrichR database names (e.g., "KEGG_2021_Human").}
}
\value{
List of data frames with results.
}
\description{
Enrichment analysis with enrichR
}
