% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_heatmap.R
\name{interactive_heatmap}
\alias{interactive_heatmap}
\title{Interactive heatmap of a gene expression}
\usage{
interactive_heatmap(expr_matrix, top_genes = 50, scale = TRUE, ...)
}
\arguments{
\item{expr_matrix}{Expression matrix (genes x samples).}

\item{top_genes}{Number of genes plus variables to display (default: 50).}

\item{scale}{If TRUE, scale rows (default: TRUE).}

\item{...}{Other parameters passed to heatmaply.}
}
\value{
htmlwidget object (viewable in RStudio or browser).
}
\description{
Create an interactive heatmap of expression data using heatmaply.
}
