% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_expression}
\alias{normalize_expression}
\title{Normalize a gene expression matrix}
\usage{
normalize_expression(expr_data, method = "log2")
}
\arguments{
\item{expr_data}{Expression matrix (genes x samples).}

\item{method}{Normalization method: "log2", "zscore", "quantile". Default: 'log2.}
}
\value{
Normalized matrix.
}
\description{
Apply different normalization methods to an expression matrix (genes x samples).
}
\examples{
data(expr_small)
norm_expr <- normalize_expression(expr_small, method = "zscore")
}
