% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_network.R
\name{plot_coexpression_network}
\alias{plot_coexpression_network}
\title{Display a co-expression network (static, custom)}
\usage{
plot_coexpression_network(g, node_anno = NULL, palette = rainbow, ...)
}
\arguments{
\item{g}{Object igraph.}

\item{node_anno}{Annotation vector/factor for nodes (optional).}

\item{palette}{Color function for groups (default: rainbow).}

\item{...}{Other parameters for plot.igraph.}
}
\description{
Display a co-expression network (static, custom)
}
