% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_network.R
\name{plot_coexpression_network_interactive}
\alias{plot_coexpression_network_interactive}
\title{Display a co-expression network interactively}
\usage{
plot_coexpression_network_interactive(g, node_anno = NULL, palette = rainbow)
}
\arguments{
\item{g}{Object igraph.}

\item{node_anno}{Annotation vector/factor for nodes (optional).}

\item{palette}{Color function for groups (default: rainbow).}
}
\description{
Display a co-expression network interactively
}
