% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Display enrichment results (barplot)}
\usage{
plot_enrichment(enrich_result, top = 10)
}
\arguments{
\item{enrich_result}{Object enrichResult (from clusterProfiler or ReactomePA).}

\item{top}{No. of terms to display (default: 10).}
}
\description{
Display enrichment results (barplot)
}
\examples{
if (requireNamespace("clusterProfiler", quietly = TRUE) &&
    requireNamespace("enrichplot", quietly = TRUE) &&
    requireNamespace("org.Hs.eg.db", quietly = TRUE)) {
  gene <- c("4312", "8318", "10874", "55143", "55388", "991", "6280", "23046", "949", "10533")
  ego <- clusterProfiler::enrichGO(
    gene          = gene,
    OrgDb         = "org.Hs.eg.db",
    keyType       = "ENTREZID",
    ont           = "BP",
    pAdjustMethod = "BH",
    pvalueCutoff  = 0.1,
    readable      = FALSE
  )
  if (nrow(ego@result) > 0) {
    plot_enrichment(ego, top = 5)
  }
}
}
