% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_pairs_interactive.R
\name{plot_gene_pair_interactive}
\alias{plot_gene_pair_interactive}
\title{Display a gene pair interactively}
\usage{
plot_gene_pair_interactive(expr_data, gene1, gene2)
}
\arguments{
\item{expr_data}{Expression matrix (genes x samples).}

\item{gene1}{First gene.}

\item{gene2}{Second gene.}
}
\description{
Use plotly for an interactive scatterplot of a gene pair.
}
\examples{
data(expr_small)
filt_expr <- filter_genes(expr_small, min_mean = 1, min_var = 0.1)
coexp <- top_gene_coexpression(filt_expr, top_n = 5)
plot_gene_pair_interactive(filt_expr, coexp$gene1[1], coexp$gene2[1])
}
